app/requirements.txt
fastapi==0.120.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-logging==3.12.1
pydantic==2.12.3
typing-extensions>=4.14.1
uvicorn==0.38.0
email-validator
